Disable Block patch option
* added the ability to disable the patch for allow/deny block place/break flags

The reason why I added this option because it breaks compatibility with some plugins that uses claims system (claiming a region for survival servers)
