+++ WGEF-Core/src/main/resources/config.yml
@@ -1,1 +1,2 @@
 update-checker: true
+disable-block-flag-patch: false
